pyproject.toml
[project]
name = "seadusloome"
version = "0.1.0"
description = "Estonian Legal Ontology Advisory Software"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-fasthtml>=0.12.0",
    "rdflib>=7.0.0",
    "pyshacl",
    "PyJWT",
    "bcrypt",
    "authlib",
    "requests",
    "psycopg[binary]>=3.2.0",
    "httpx",
    "python-docx",
    "anthropic>=0.93.0",
    "voyageai>=0.3.7",
    "sentry-sdk[starlette]>=2.57.0",
]

[dependency-groups]
dev = [
    "ruff",
    "pyright",
    "pytest",
    "vcrpy>=8.1.1",
    "pytest-recording>=0.13.4",
]

[tool.ruff]
target-version = "py313"
line-length = 99

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.ruff.lint.per-file-ignores]
"app/**/*.py" = ["F403", "F405"]
# UI component functions follow FastHTML's PascalCase convention for FT builders
"app/ui/**/*.py" = ["F403", "F405", "N802"]

[tool.pyright]
pythonVersion = "3.13"
typeCheckingMode = "basic"
reportMissingTypeStubs = false
reportWildcardImportFromLibrary = false
